Extract NodeId type into a submodule
